combined serializer package into one single module - still wip
